Move polynomial_encode_decode into its own file
