+++ drake/systems/controllers/BUILD.bazel
@@ -158,6 +158,19 @@ drake_cc_library(
     ],
 )
 
+drake_cc_library(
+    name = "polynomial_encode_decode",
+    srcs = ["polynomial_encode_decode.cc"],
+    hdrs = ["polynomial_encode_decode.h"],
+    deps = [
+        "//drake/common/trajectories:piecewise_polynomial",
+        "//drake/lcmtypes:piecewise_polynomial",
+        "//drake/lcmtypes:polynomial",
+        "//drake/lcmtypes:polynomial_matrix",
+        "//drake/util",
+    ],
+)
+
 # This library is empty unless Gurobi is available.
 drake_cc_library(
     name = "instantaneous_qp_controller",
@@ -171,6 +184,7 @@ drake_cc_library(
     }),
     deps = [
         ":control_util",
+        ":polynomial_encode_decode",
         ":qp_common",
         ":yaml_util",
         "//drake/common:essential",
@@ -187,6 +201,8 @@ drake_cc_library(
     ],
 )
 
+# === test/ ===
+
 drake_cc_library(
     name = "zmp_test_util",
     testonly = 1,
@@ -197,7 +213,14 @@ drake_cc_library(
     ],
 )
 
-# === test/ ===
+drake_cc_googletest(
+    name = "polynomial_encode_decode_test",
+    deps = [
+        ":polynomial_encode_decode",
+        "//drake/common/test_utilities",
+        "//drake/common/trajectories:random_piecewise_polynomial",
+    ],
+)
 
 drake_cc_googletest(
     name = "inverse_dynamics_test",
