+++ drake/systems/controllers/InstantaneousQPController.cpp
@@ -20,6 +20,7 @@
 #include "drake/multibody/parsers/urdf_parser.h"
 #include "drake/solvers/fast_qp.h"
 #include "drake/systems/controllers/controlUtil.h"
+#include "drake/systems/controllers/polynomial_encode_decode.h"
 #include "drake/systems/controllers/yaml_util.h"
 #include "drake/util/drakeGeometryUtil.h"
 #include "drake/util/lcmUtil.h"
